Update message support and template

+++ application/modules/master/controllers/addcustomer.php
@@ -57,6 +57,18 @@ class addcustomer extends CI_Controller {
 		ini_set('display_errors','off'); 				
 		$this->load->model('adminheader_model','top_model');
     }
+	/** Admin always; sub-admin only when Roles & Responsibilities Delete Customer is granted. **/
+	protected function can_delete_customer() {
+		if ($this->session->userdata('usertype') == 'admin') {
+			return true;
+		}
+		$roleResponsible = $this->top_model->get_responsibilities();
+		return (
+			is_array($roleResponsible)
+			&& array_key_exists('delete_customer', $roleResponsible)
+			&& (string) $roleResponsible['delete_customer'] === '1'
+		);
+	}
 	public function index(){ 		 //*****  View Loading  *****//
 		if($this->session->userdata('usertype') == 'subadmin'){
 			$this->head['roleResponsible'] = $this->top_model->get_responsibilities();
@@ -70,6 +82,7 @@ class addcustomer extends CI_Controller {
 		// No longer loading all records - using server-side pagination instead
 		$data['record'] = array();
 		$data['zone'] = $this->my_model->get_zone();
+		$data['can_delete_customer'] = $this->can_delete_customer();
 		//$header['host'] = $this->comm_model->get_single_record();				
 		//$header['record_info'] = $this->top_model->get_last_login_details(1);
 		$this->load->view($this->headerPage,$this->head);
@@ -142,6 +155,7 @@ class addcustomer extends CI_Controller {
 			$records = $this->my_model->get_paginated_records($start, $length, $search, $order_column, $order_dir, $zone);
 			$total_records = $this->my_model->get_total_count('', $zone);
 			$filtered_records = $this->my_model->get_total_count($search, $zone);
+			$can_delete_customer = $this->can_delete_customer();
 			
 			// Format data for DataTables
 			$data = array();
@@ -152,74 +166,46 @@ class addcustomer extends CI_Controller {
 				$row_id = isset($row['id']) ? $row['id'] : 0;
 				
 				if($row_status == 1) {
-					$status_html = '<span class="label label-success arrowed-in arrowed-in-right"><a href="JavaScript:if(confirm(\'Are you sure want to Chanage the Status?\')==true){window.location=\''.ADMIN_URL.'addcustomer/status/'.$row_id.'/'.$row_status.'\';}" style="color:#FFF; text-decoration:none;">Active</a></span>';
+					$status_html = '<a href="JavaScript:if(confirm(\'Are you sure want to Change the Status?\')==true){window.location=\''.ADMIN_URL.'addcustomer/status/'.$row_id.'/'.$row_status.'\';}" class="badge badge-success badge-pill">Active</a>';
 				} elseif($row_status == 0) {
-					$status_html = '<span class="label label-danger arrowed"><a href="JavaScript:if(confirm(\'Are you sure want to Chanage the Status?\')==true){window.location=\''.ADMIN_URL.'addcustomer/status/'.$row_id.'/'.$row_status.'\';}" style="color:#FFF; text-decoration:none;">De-Active</a></span>';
+					$status_html = '<a href="JavaScript:if(confirm(\'Are you sure want to Change the Status?\')==true){window.location=\''.ADMIN_URL.'addcustomer/status/'.$row_id.'/'.$row_status.'\';}" class="badge badge-warning badge-pill">De-Active</a>';
 				} elseif($row_status == 2) {
-					$status_html = '<span class="label label-danger arrowed"><a href="JavaScript:if(confirm(\'Are you sure want to Chanage the Status?\')==true){window.location=\''.ADMIN_URL.'addcustomer/status/'.$row_id.'/'.$row_status.'\';}" style="color:#FFF; text-decoration:none;">Disconnected</a></span>';
+					$status_html = '<a href="JavaScript:if(confirm(\'Are you sure want to Change the Status?\')==true){window.location=\''.ADMIN_URL.'addcustomer/status/'.$row_id.'/'.$row_status.'\';}" class="badge badge-danger badge-pill">Disconnected</a>';
 				}
 				
+				$delete_btn_html = '';
+				if ($can_delete_customer) {
+					$delete_btn_html = '<a class="btn btn-outline-danger sa4-confirm-delete" href="javascript:void(0);" data-sa4-delete-url="'.ADMIN_URL.'addcustomer/delete/'.$row_id.'" title="Delete" data-toggle="tooltip">
+										<i class="fal fa-times"></i>
+									</a>';
+				}
+
 				$action_html = '<input type="hidden" name="id_'.$i.'" id="id_'.$i.'" value="'.$row_id.'">
 								<input type="hidden" name="customerid_'.$i.'" id="customerid_'.$i.'" value="'.(isset($row['customer_id']) ? $row['customer_id'] : '').'">
 								<input type="hidden" name="billingplansid_'.$i.'" id="billingplansid_'.$i.'" value="'.(isset($row['billingplans']) ? $row['billingplans'] : '').'">
-								<div class="visible-md visible-lg hidden-sm hidden-xs action-buttons">
-									<a class="blue view-customer-btn" href="javascript:void(0);" data-customer-id="'.$row_id.'" title="View">
-										<i class="fa fa-info-circle"></i>
-									</a>	
-									<a class="green edit-customer-btn" href="javascript:void(0);" data-customer-id="'.$row_id.'" title="Edit">
-										<i class="fa fa-edit"></i>
+								<div class="btn-group btn-group-sm" role="group">
+									<a class="btn btn-outline-primary view-customer-btn" href="javascript:void(0);" data-customer-id="'.$row_id.'" title="View" data-toggle="tooltip">
+										<i class="fal fa-eye"></i>
 									</a>
-									<a class="orange set-password-btn" href="javascript:void(0);" data-customer-id="'.$row_id.'" data-customer-code="'.(isset($row['customer_id']) ? $row['customer_id'] : '').'" title="Set Login Password">
-										<i class="fa fa-key"></i>
+									<a class="btn btn-outline-success edit-customer-btn" href="javascript:void(0);" data-customer-id="'.$row_id.'" title="Edit" data-toggle="tooltip">
+										<i class="fal fa-edit"></i>
 									</a>
-									<a class="red" href="JavaScript:if(confirm(\'Confirm Delete?\')==true){window.location=\''.ADMIN_URL.'addcustomer/delete/'.$row_id.'\';}" title="Delete">
-										<i class="fa fa-remove"></i>
+									<a class="btn btn-outline-warning set-password-btn" href="javascript:void(0);" data-customer-id="'.$row_id.'" data-customer-code="'.(isset($row['customer_id']) ? htmlspecialchars($row['customer_id'], ENT_QUOTES, 'UTF-8') : '').'" title="Set Login Password" data-toggle="tooltip">
+										<i class="fal fa-key"></i>
 									</a>
-								</div>
-								<div class="visible-xs visible-sm hidden-md hidden-lg">
-									<div class="inline position-relative">
-										<button class="btn btn-minier btn-yellow dropdown-toggle" data-toggle="dropdown">
-											<i class="icon-caret-down icon-only bigger-120"></i>
-										</button>
-										<ul class="dropdown-menu dropdown-only-icon dropdown-yellow pull-right dropdown-caret dropdown-close">
-											<li>
-												<a href="javascript:void(0);" class="edit-customer-btn" data-customer-id="'.$row_id.'" data-rel="tooltip" title="Edit">
-													<span class="green">
-														<img src="'.base_url().'images/favicon/document-edit.gif">
-													</span>
-												</a>
-											</li>
-											<li>
-												<a class="blue view-customer-btn" href="javascript:void(0);" data-customer-id="'.$row_id.'" data-rel="tooltip" title="View">
-													<span class="blue">
-														<img src="'.base_url().'images/favicon/view_icon.gif">
-													</span>
-												</a>			
-											</li>
-											<li>
-												<a href="javascript:void(0);" class="set-password-btn" data-customer-id="'.$row_id.'" data-customer-code="'.(isset($row['customer_id']) ? $row['customer_id'] : '').'" data-rel="tooltip" title="Set Login Password">
-													<span class="orange">
-														<i class="fa fa-key"></i>
-													</span>
-												</a>
-											</li>
-											<li>
-												<a href="JavaScript:if(confirm(\'Confirm Delete?\')==true){window.location=\''.ADMIN_URL.'addcustomer/delete/'.$row_id.'\';}" class="tooltip-error" data-rel="tooltip" title="Delete">
-													<span class="red">
-														<img src="'.base_url().'images/favicon/delete.png">
-													</span>
-												</a>
-											</li>
-										</ul>
-									</div>
+									'.$delete_btn_html.'
 								</div>';
+
+				$checkbox_html = $can_delete_customer
+					? '<input type="checkbox" class="ace" name="delete_ids[]" id="delete_ids[]" value="'.$row_id.'" />'
+					: '';
 				
 				$data[] = array(
-					'<input type="checkbox" class="ace" name="delete_ids[]" id="delete_ids[]" value="'.$row_id.'" />',
+					$checkbox_html,
 					$i++,
 					stripslashes(isset($row['customer_id']) ? $row['customer_id'] : ''),
 					stripslashes((isset($row['last_name']) ? $row['last_name'] : '').',  '.(isset($row['first_name']) ? $row['first_name'] : '').'  '.(isset($row['middle_name']) ? $row['middle_name'] : '')),
-					'<span style="width:25%;">'.stripslashes(isset($row['address']) ? $row['address'] : '').'</span>',
+					stripslashes(isset($row['address']) ? $row['address'] : ''),
 					stripslashes(isset($row['meter_number']) ? $row['meter_number'] : ''),
 					stripslashes(isset($row['zones']) ? $row['zones'] : ''),
 					stripslashes(isset($row['classification_name']) ? $row['classification_name'] : ''),
@@ -445,57 +431,66 @@ class addcustomer extends CI_Controller {
 	}
 	public function adminconfigurationupdate()
 	{
-		// Ensure database connection is available
 		if (!isset($this->db) || !is_object($this->db)) {
 			$this->load->database();
 		}
-		$this->load->library('image_lib');
-		$adminid						= $this->input->post('adminid');
-		$data['name'] 					= $this->input->post('name');
-		$data['email'] 					= $this->input->post('email');
-		$data['established'] 			= date ("Y-m-d", strtotime($this->input->post('established')));
-		$data['contact1']	 			= $this->input->post('contact1');
-		$data['contactperson'] 			= $this->input->post('contactperson');
-		$data['contactpersonphone'] 	= $this->input->post('contactperson');
-		$data['website']			 	= $this->input->post('website');
-		$data['address1']			 	= $this->input->post('address1');
-		$data['about']				 	= $this->input->post('about');
-		$result							= $this->my_model->get_adminrecord_update($adminid,$data);
-						$config = array(
-										'upload_path'   => './images/logo',
-										'allowed_types' => 'gif|jpg|png',
-										/*'max_size'      => '10000',
-										'max_width'     => '1024',
-										'max_height'    => '768',*/
-										'encrypt_name'  => false,
-									   );
-						$this->load->library('upload', $config);
-						$this->image_lib->resize();
-			if($result){
-						if ($this->upload->do_upload('userfile')) 
-							{	
-							   	$upload_data = $this->upload->data();
-									$data_ary = array(
-														'title'     => $upload_data['client_name'],
-														'file'      => $upload_data['file_name'],
-														'width'     => $upload_data['image_width'],
-														'height'    => $upload_data['image_height'],
-														'type'      => $upload_data['image_type'],
-														'size'      => $upload_data['file_size'],
-														'date'      => date('Y-m-d')
-													  );
-														$this->db->where('adminid', $adminid);
-														$this->db->update('tbl_adminlogo', $data_ary);
-							}
-		
-						$data['record'] = $this->my_model->get_adminrecord_edit($adminid);
-						$this->load->view($this->headerPage,$this->head);
-						$this->load->view(adminconfiguration,$data);
-					}
-			else
-			{
-				$data['msg'] = "Not Updated...";
-			}	
+
+		$adminid = (int) $this->input->post('adminid');
+		if ($adminid <= 0) {
+			$this->session->set_flashdata('msg_err', 'Invalid admin configuration record.');
+			redirect('master/addcustomer/adminconfiguration');
+			return;
+		}
+
+		$established_raw = trim((string) $this->input->post('established'));
+		$established = $established_raw !== '' ? date('Y-m-d', strtotime($established_raw)) : null;
+
+		$update = array(
+			'name'               => trim((string) $this->input->post('name')),
+			'email'              => trim((string) $this->input->post('email')),
+			'established'        => $established,
+			'contact1'           => trim((string) $this->input->post('contact1')),
+			'contactperson'      => trim((string) $this->input->post('contactperson')),
+			'contactpersonphone' => trim((string) $this->input->post('contactpersonphone')),
+			'website'            => trim((string) $this->input->post('website')),
+			'address1'           => trim((string) $this->input->post('address1')),
+			'about'              => trim((string) $this->input->post('about')),
+		);
+
+		$result = $this->my_model->get_adminrecord_update($adminid, $update);
+
+		if (!$result) {
+			$this->session->set_flashdata('msg_err', 'Not Updated...');
+			redirect('master/addcustomer/editadminconfiguration/' . $adminid);
+			return;
+		}
+
+		if (!empty($_FILES['userfile']['name'])) {
+			$this->load->library('image_lib');
+			$config = array(
+				'upload_path'   => './images/logo',
+				'allowed_types' => 'gif|jpg|png|jpeg',
+				'encrypt_name'  => false,
+			);
+			$this->load->library('upload', $config);
+			if ($this->upload->do_upload('userfile')) {
+				$upload_data = $this->upload->data();
+				$logo_data = array(
+					'title'  => $upload_data['client_name'],
+					'file'   => $upload_data['file_name'],
+					'width'  => $upload_data['image_width'],
+					'height' => $upload_data['image_height'],
+					'type'   => $upload_data['image_type'],
+					'size'   => $upload_data['file_size'],
+					'date'   => date('Y-m-d'),
+				);
+				$this->db->where('adminid', $adminid);
+				$this->db->update('tbl_adminlogo', $logo_data);
+			}
+		}
+
+		$this->session->set_flashdata('msg_succ', 'Admin configuration updated successfully.');
+		redirect('master/addcustomer/adminconfiguration');
 	}
 	/** Save Customer Login Password **/
 	public function save_customer_password() {
@@ -1006,6 +1001,11 @@ class addcustomer extends CI_Controller {
 	/** Delete Function **/
 	public function delete($id){ 
 		$data['msg'] ='';
+		if (!$this->can_delete_customer()) {
+			$this->session->set_flashdata('msg_succ', 'You are not authorized to delete customers.');
+			redirect($this->listPage_redirect);
+			return;
+		}
 		if($id){
 			$result = $this->my_model->delete_record($id);
 			if($result){
@@ -1020,6 +1020,11 @@ class addcustomer extends CI_Controller {
 	/** Multiple Delete Function **/
 	public function multi_delete(){
 		$data['msg'] ='';
+		if (!$this->can_delete_customer()) {
+			$this->session->set_flashdata('msg_succ', 'You are not authorized to delete customers.');
+			redirect($this->listPage_redirect);
+			return;
+		}
 		if($this->input->post('delete_ids') != ''){
 			$delete_ids = $this->input->post('delete_ids');
 			for($i=0;$i<count($delete_ids);$i++){
